+++ src/include/Analyze/AnalyzeCli.hpp
@@ -5,4 +5,5 @@
 #include "argparse/argparse.hpp"
 
 std::vector<std::filesystem::path> getFilesToAnalyze(const std::vector<std::string>& paths, const std::vector<std::string>& ignoreGlobPatterns);
+std::unordered_map<std::string, std::filesystem::path> processDefinitionsFilePaths(const argparse::ArgumentParser& program);
 int startAnalyze(const argparse::ArgumentParser& program);