+++ tests/Diagnostics.test.cpp
@@ -12,7 +12,7 @@ TEST_CASE_FIXTURE(Fixture, "document_diagnostics_sends_information_for_required_
     client->capabilities.textDocument->diagnostic->relatedDocumentSupport = true;
 
     // Don't show diagnostic for game indexing
-    loadDefinition("declare game: any");
+    loadDefinition("@extra", "declare game: any");
 
     registerDocumentForVirtualPath(newDocument("required.luau", R"(
         --!strict
@@ -36,7 +36,7 @@ TEST_CASE_FIXTURE(Fixture, "document_diagnostics_does_not_send_information_for_r
     client->capabilities.textDocument->diagnostic->relatedDocumentSupport = false;
 
     // Don't show diagnostic for game indexing
-    loadDefinition("declare game: any");
+    loadDefinition("@extra", "declare game: any");
 
     registerDocumentForVirtualPath(newDocument("required.luau", R"(
         --!strict